selesai section 23

+++ app/Http/Controllers/PostController.php
@@ -35,7 +35,7 @@ class PostController extends Controller
         // menggunakan cache
         // $posts = Cache::remember('posts', 60 , function(){ /*durasi penyimpanan dalam detik*/
         // $posts = Cache::rememberForever('posts' , function(){ /*cache selamanya*/
-            $posts = Cache::remember('posts-page-'.request('page',1), 5 , function(){ /*untuk menangani cache dengan pagination*/
+            $posts = Cache::remember('posts-page-'.request('page',1), 10 , function(){ /*untuk menangani cache dengan pagination*/
             return Post::with('category')->paginate(3);
         });
         return view('layouts.index', compact('posts'));
@@ -44,11 +44,17 @@ class PostController extends Controller
 
     public function create()
     {
+
+        //untuk cek otoritasi
+        // $this->authorize('create_post'); /*menggunakan gate saja*/
+        $this->authorize('create', Post::class);/*menggunakan policy*/
         $categories = Category::all();
         return view('layouts.create', compact('categories'));
     }
     public function store(Request $request)
     {
+        // $this->authorize('create_post');
+        $this->authorize('create', Post::class);/*menggunakan policy*/
         $request->validate([
             'image' => ['required', 'max:2028', 'image'],
             'title' => ['required', ' max:255' ],
@@ -74,9 +80,12 @@ class PostController extends Controller
         return view('layouts.show',compact('post'));
     }
 
-    public function edit($id)
-    {
+    public function edit($id)    {
+
+        // $this->authorize('edit_post');
         $post = Post::findOrFail($id);
+
+        $this->authorize('update', $post);/*menggunakan policy*/
         $categories = Category::all();
         return view('layouts.edit', compact('post' , 'categories') );
     }
@@ -84,12 +93,16 @@ class PostController extends Controller
     public function update(Request $request, $id)
     {
         // return $request;
+        $post = Post::findOrFail($id);
+
         $request->validate([
             'title' => ['required', ' max:255' ],
             'category_id' => ['required', 'integer'],
             'description' => ['required']
         ]);
-        $post = Post::findOrFail($id);
+
+
+        $this->authorize('update', $post);/*menggunakan policy*/
         if($request->hasFile('image')){
             $request->validate([
                 'image' => ['required', 'max:2028', 'image'], ]);
@@ -106,34 +119,35 @@ class PostController extends Controller
         return redirect()->route('posts.index');
     }
 
-            /**
-             * Remove the specified resource from storage.
-             */
-    public function destroy($id)
-    {
+    public function destroy($id)    {
+    // $this->authorize('delete_post');
     $post = Post::findOrFail($id);
+
+    $this->authorize('delete',$post);
+
     $post->delete();
-    return redirect()->route('posts.index');
-    }
+    return redirect()->route('posts.index');    }
 
     public function trashed(){
+        $this->authorize('delete_post');
         $posts = Post::onlyTrashed()->get();
-        return view('layouts.trashed',compact('posts'));
-    }
+        return view('layouts.trashed',compact('posts'));}
 
     public function restore($id){
-
+        $this->authorize('delete_post');
         $post = Post::onlyTrashed()->findOrFail($id);
         $post->restore();
-        return redirect()->route('posts.index');
-    }
+        return redirect()->route('posts.index');}
 
     public function forceDelete($id){
+    $this->authorize('delete_post');
+    $post = Post::onlyTrashed()->findOrFail($id);
+    File::delete(public_path($post->image));
+    $post->forceDelete();
+    return redirect()->route('posts.index');}
 
-        $post = Post::onlyTrashed()->findOrFail($id);
-        File::delete(public_path($post->image));
-        $post->forceDelete();
 
-    return redirect()->route('posts.index');
-    }
+
+
+
     }
